Store BG tile attribute shift registers as 16-bit integers

+++ src/ppu.h
@@ -108,7 +108,7 @@ typedef struct {
 	/* TODO: make everything below internal */
 	uint8_t spr_in_range, ovr_check_done, oamend;
 
-	uint8_t bg_attr1, bg_attr2;
+	uint16_t bg_attr1, bg_attr2;
 	uint16_t bg_bmp1, bg_bmp2;
 
 	uint8_t soam[32]; /* Secondary OAM */
